requirements.txt
asgiref==3.9.1
attrs==25.4.0
branca==0.8.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==46.0.1
diff-match-patch==20241021
Django==6.0
django-allauth==65.11.2
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-htmx==1.25.0
django-import-export==4.4.0
django-silk==5.5.0
django-unfold==0.74.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.5.1
fido2==2.0.0
folium==0.20.0
gprof2dot==2025.4.14
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
numpy==2.4.3
oauthlib==3.3.1
pandas==3.0.1
pillow==11.3.0
promise==2.3
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xyzservices==2025.11.0
